embed videos everywhere (#1707)
* create ProposalDescription component

* add to feed, competitions and submissions

* help text for competition+submission description
