+++ src/components/Proposal/ProposalDetails.scss
@@ -310,10 +310,6 @@ button.disabled {
   }
 }
 
-iframe.embeddedVideo {
-  margin-top: 10px;
-}
-
 @media only screen and (max-width: 425px) {
   .proposal {
     display: block;
