+++ src/components/Scheme/ContributionRewardExtRewarders/Competition/SubmissionDetails.tsx
@@ -17,8 +17,7 @@ import { DiscussionEmbed } from "disqus-react";
 import { RouteComponentProps } from "react-router-dom";
 import { getSubmission, getSubmissionVoterHasVoted, getCompetitionVotes, CompetitionStatus } from "./utils";
 import * as css from "./Competitions.scss";
-
-const ReactMarkdown = require("react-markdown");
+import ProposalDescription from "components/Shared/ProposalDescription";
 
 type ISubscriptionState = [ICompetitionSuggestionState, boolean, Array<CompetitionVote>];
 
@@ -129,11 +128,7 @@ class SubmissionDetails extends React.Component<IProps, null> {
 
         { submission.description ?
           <div className={css.description}>
-            <ReactMarkdown source={submission.description}
-              renderers={{link: (props: { href: string; children: React.ReactNode }) => {
-                return <a href={props.href} target="_blank" rel="noopener noreferrer">{props.children}</a>;
-              }}}
-            />
+            <ProposalDescription description={submission.description} />
           </div>
           : "" }
